// packages/muya/src/inlineRenderer/types.ts
export interface IHighlight {
    start: number;
    end: number;
    active: boolean | undefined;
}

export type Labels = Map<
    string,
    {
        href: string;
        title: string;
    }
>;

export type Rules = Record<string, RegExp>;

export interface ITokenizerFacOptions {
    superSubScript: boolean;
    footnote: boolean;
}

export interface ITokenizerOptions {
    highlights?: IHighlight[];
    hasBeginRules?: boolean;
    labels?: Labels;
    options?: ITokenizerFacOptions;
}

export interface ITokenRange {
    start: number;
    end: number;
}

export type Token
    = | BeginRuleToken
        | ReferenceDefinitionToken
        | TextToken
        | BacklashToken
        | StrongEmToken
        | CodeEmojiMathToken
        | DelToken
        | SuperSubScriptToken
        | FootnoteIdentifierToken
        | ImageToken
        | LinkToken
        | ReferenceLinkToken
        | ReferenceImageToken
        | HTMLEscapeToken
        | AutoLinkExtensionToken
        | AutoLinkToken
        | HTMLTagToken
        | SoftLineBreakToken
        | HardLineBreakToken
        | TailHeaderToken;

export interface IBaseToken {
    raw: string;
    parent: Token[];
    range: ITokenRange;
    highlights?: IHighlight[];
}
// ...
export type CodeEmojiMathToken = IBaseToken & {
    type: 'inline_code' | 'emoji' | 'inline_math';
    marker: string;
    content: string;
    backlash: string;
    displayMode?: boolean;
};